pipeline/requirements.txt
pandas
PyMySQL
python-dotenv
cryptography
pyarrow
awswrangler
boto3